pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "cz-benchmarks"
version = "0.10.0"
description = "A framework for benchmarking single-cell machine learning models"
authors = [{ name = "Chan Zuckerberg Initiative", email = "pypi@chanzuckerberg.com" }]
readme = "README-pypi.md"
license = { text = "MIT" }
requires-python = ">=3.10"
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Operating System :: OS Independent",
    "License :: OSI Approved :: MIT License",
]
dependencies = [
    # Core dependencies
    "docker>=6.1.0",
    "pyyaml>=6.0",
    "boto3>=1.28.0",
    # Data handling
    "anndata>=0.9.0",
    "h5py>=3.8.0",
    "dill>=0.3.6",
    "scikit-learn",
    "scanpy",
    "igraph>=0.11.8",
    "leidenalg>=0.10.2",
    "hydra-core>=1.3.2",
    "scib-metrics>=0.5.1",
    "scikit-misc>=0.5.1",
    "typing-extensions>=4.13.0",
    "pydantic>=2.0.0",
    "hnswlib>=0.8.0",
    "tomli>=2.2.1",
    "boto3-stubs-lite[s3]>=1.38.0",
    "typer>=0.16.0",
]

[project.optional-dependencies]
interactive = [
    "jupyterlab>=4.0.0",
]

build = [
    "build>=1.2.2.post1",
    "wheel>=0.45.1",
    "setuptools>=78.1.0",
    "twine>=6.1.0",
]

dev = [
    "boto3-stubs>=1.37.26",
    "botocore-stubs>=1.37.26",
    "ruff>=0.11.2",
    "mypy>=1.15.0",
    "pandas-stubs>=2.2.3.250308",
    "pre-commit>=4.2.0",
    "pytest>=8.3.5",
    "pytest-cov>=6.0.0",
    "pytest-mock>=3.14.0",
    "types-pyyaml>=6.0.12.20250402",
]

docs = [
    "toml",
    "sphinx==8.1.3",
    "sphinx-autoapi",
    "sphinx-autodoc-typehints",
    "sphinx-markdown-builder",
    "myst-parser",
    "sphinxcontrib.mermaid", 
    "sphinx-rtd-theme",
    "sphinx-copybutton",
    "nbsphinx",
    "linkify-it-py",
    
]

[project.urls]
Homepage = "https://github.com/chanzuckerberg/cz-benchmarks"
Repository = "https://github.com/chanzuckerberg/cz-benchmarks"

[tool.hatch.build.targets.wheel]
packages = ["src/czbenchmarks"]

[project.scripts]
czbenchmarks = "czbenchmarks.cli.cli:main"

[tool.pytest.ini_options]
markers = "integration: marks tests as integration tests (skipped by default, run with `-m integration`)"
addopts = "-m 'not integration'"

[tool.coverage.report]
omit = [
    "src/czbenchmarks/cli/*"
]